AC_lesson_4 with corrections

+++ src/main/java/student_andrejs_cekalins/lesson_4/level_7/Task_24/Calculator.java
@@ -1,19 +1,19 @@
 package student_andrejs_cekalins.lesson_4.level_7.Task_24;
 
 class Calculator {
-    public int sum(int firstNumber, int secondNumber) {
+    public double sum(double firstNumber, double secondNumber) {
         return firstNumber + secondNumber;
     }
 
-    public int sub(int firstNumber, int secondNumber) {
+    public double sub(double firstNumber, double secondNumber) {
         return firstNumber - secondNumber;
     }
 
-    public int mul(int firstNumber, int secondNumber) {
+    public double mul(double firstNumber, double secondNumber) {
         return firstNumber * secondNumber;
     }
 
-    public int div(int firstNumber, int secondNumber) {
+    public double div(double firstNumber, double secondNumber) {
         return firstNumber / secondNumber;
     }
